chore: useCollapse in place of SideBarCollapser
