+++ packages/fuselage/src/components/SideBarV2/hooks/useCollapse.tsx
@@ -1,0 +1,66 @@
+import { useToggle, useUniqueId } from '@rocket.chat/fuselage-hooks';
+import type { KeyboardEventHandler, MouseEvent } from 'react';
+
+export const useCollapse = ({
+  expanded: propExpanded,
+  defaultExpanded,
+  disabled,
+  noncollapsible,
+  tabIndex = 0,
+}: {
+  expanded?: boolean;
+  defaultExpanded?: boolean;
+  disabled?: boolean;
+  noncollapsible?: boolean;
+  tabIndex?: number;
+}) => {
+  const [stateExpanded, toggleStateExpanded] = useToggle(defaultExpanded);
+  const expanded = propExpanded || stateExpanded;
+
+  const panelExpanded = noncollapsible || expanded;
+
+  const titleId = useUniqueId();
+  const panelId = useUniqueId();
+
+  const handleClick = (e: MouseEvent<HTMLElement>) => {
+    if (disabled) {
+      return;
+    }
+    e.currentTarget?.blur();
+    toggleStateExpanded();
+  };
+
+  const handleKeyDown: KeyboardEventHandler = (event) => {
+    if (disabled || event.currentTarget !== event.target) {
+      return;
+    }
+
+    if ([13, 32].includes(event.keyCode)) {
+      event.preventDefault();
+
+      if (event.repeat) {
+        return;
+      }
+
+      toggleStateExpanded();
+    }
+  };
+
+  const collapsibleProps = {
+    'aria-controls': panelId,
+    'aria-expanded': expanded ? 'true' : 'false',
+    'tabIndex': !disabled ? tabIndex : undefined,
+    'onClick': handleClick,
+    'onKeyDown': handleKeyDown,
+  } as const;
+
+  const nonCollapsibleProps = {
+    'aria-disabled': 'true',
+    'aria-expanded': 'true',
+    'aria-labelledby': titleId,
+  } as const;
+
+  const barProps = noncollapsible ? nonCollapsibleProps : collapsibleProps;
+
+  return { barProps, titleId, panelId, panelExpanded, expanded };
+};
